refactor(face-core): own the Face Package schema; render re-exports it
Type-home move per docs/redesign/06-track-2-implementation.md §3.4(a):
the VIZIJ_bundle payload types (VizijBundleExtension, graph/pose/
animation entries, VizijSpeechConfig, ...) are L1 vocabulary, so they
move verbatim from @vizij/render/src/types/vizij-bundle.ts to
@vizij/face-core/src/facePackage.ts.

- render's types/vizij-bundle.ts becomes an explicit re-export shim —
  every previously exported name is preserved, so render's public API
  is unchanged (api snapshots updated; render's rolled-up d.ts now
  references face-core instead of inlining the schema).
- runtime-react imports the schema from face-core; src/core/ no longer
  imports anything from @vizij/render, which keeps the extracted
  FaceRuntime layer honest ahead of the R1 package move (PR-3 will not
  need a face-core -> render types edge, avoiding the layering cycle).
- render + runtime-react gain a workspace dep on face-core (fixed
  release line already groups the trio).

Verified: face-core/render/runtime-react build + typecheck green;
runtime-react 83 tests + headless Node smoke pass; vizij-authoring
typecheck + 729 tests green against the rebuilt packages; api:check OK.
