+++ packages/@vizij/render/package.json
@@ -39,6 +39,7 @@
     "prepack": "pnpm run build && pnpm run test && pnpm run typecheck && pnpm run lint && pnpm run size"
   },
   "dependencies": {
+    "@vizij/face-core": "workspace:*",
     "@vizij/utils": "workspace:*",
     "clsx": "^2.1.1",
     "deep-equal": "^2.2.3",
