+++ packages/@vizij/runtime-react/package.json
@@ -40,6 +40,7 @@
   },
   "dependencies": {
     "@vizij/animation-module": "^0.2.0",
+    "@vizij/face-core": "workspace:*",
     "@vizij/node-graph-authoring": "workspace:*",
     "@vizij/render": "workspace:*",
     "@vizij/runtime": "^1.1.0",
